+++ tests/templates/kuttl/mount-dags-gitsync/dag_metrics.py
@@ -7,10 +7,12 @@ import logging
 
 
 def assert_metric(role, metric):
-    metric_response = requests.get(f"http://airflow-{role}-default:9102/metrics")
-    assert (
-        metric_response.status_code == 200
-    ), f"Metrics could not be retrieved from the {role}."
+    metric_response = requests.get(
+        f"http://airflow-{role}-default-metrics:9102/metrics"
+    )
+    assert metric_response.status_code == 200, (
+        f"Metrics could not be retrieved from the {role}."
+    )
     return metric in metric_response.text